+++ src/game/states/save_game_menu.hpp
@@ -11,7 +11,7 @@ namespace Game {
             SaveGameMenu();
             ~SaveGameMenu();
             void HandleEvent(const SDL_Event& event);
-            Transition Update(const double deltaTime);
+            Transition Update(const float deltaTime);
             void Render();
             void SaveGame(const unsigned int slot);
